requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.4
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
kombu==5.3.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==7.7.0
typing-extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
